fixing the socket test.
