+++ .github/workflows/test.yaml
@@ -20,7 +20,7 @@ jobs:
         image: redis
         ports:
           - 6379:6379
-        options: --entrypoint redis-server
+        options: -v /tmp/redis:/tmp/redis --entrypoint redis-server --unixsocket /tmp/redis/redis.sock
 
     steps:
       - uses: actions/checkout@v2
