+++ tests/cli_tests/test_dsn.py
@@ -34,7 +34,7 @@ def test_using_dsn_unix():
     config_content = dedent(
         """
         [alias_dsn]
-        unix = unix:///tmp/redis.sock?db=3
+        unix = unix:///tmp/redis/redis.sock?db=3
         """
     )
     with open("/tmp/iredisrc", "w+") as etc_config:
@@ -42,6 +42,6 @@ def test_using_dsn_unix():
 
     cli = pexpect.spawn("iredis --iredisrc /tmp/iredisrc --dsn unix", timeout=2)
     cli.logfile_read = open("cli_test.log", "ab")
-    cli.expect(["iredis", "redis /tmp/redis.sock[3]>"])
+    cli.expect(["iredis", "redis /tmp/redis/redis.sock[3]>"])
 
     cli.close()
